+++ src/common/components/modals/CreateNewEventModal/useCreateEventConditions.ts
